update code to reflect changes to ComplexDoubleType
- one change is the use of doubles rather than floats
- also replaced div(1,z) with recip(z) after determining it is faster

+++ imglib2/core/src/main/java/net/imglib2/type/numeric/complex/ComplexDoubleType.java
@@ -54,7 +54,7 @@ public class ComplexDoubleType extends AbstractComplexType<ComplexDoubleType> im
 	}
 	
 	// this is the constructor if you want it to be a variable
-	public ComplexDoubleType( final float r, final float i )
+	public ComplexDoubleType( final double r, final double i )
 	{
 		img = null;
 		dataAccess = new DoubleArray( 2 );
@@ -103,7 +103,7 @@ public class ComplexDoubleType extends AbstractComplexType<ComplexDoubleType> im
 	@Override
 	public void setImaginary( final double i ){ dataAccess.setValue( imaginaryI, i ); }
 	
-	public void set( final float r, final float i ) 
+	public void set( final double r, final double i ) 
 	{ 
 		dataAccess.setValue( realI, r );
 		dataAccess.setValue( imaginaryI, i );
